refactor: [M3-9367] - Move betas queries
