+++ packages/queries/src/betas/betas.ts
@@ -1,6 +1,6 @@
-import { getBeta, getBetas } from '@linode/api-v4/lib/betas';
-import { createQueryKeys } from '@lukemorales/query-key-factory';
 import { keepPreviousData, useQuery } from '@tanstack/react-query';
+
+import { betaQueries } from './keys';
 
 import type { Beta } from '@linode/api-v4/lib/betas';
 import type {
@@ -9,17 +9,6 @@ import type {
   Params,
   ResourcePage,
 } from '@linode/api-v4/lib/types';
-
-export const betaQueries = createQueryKeys('betas', {
-  beta: (id: string) => ({
-    queryFn: () => getBeta(id),
-    queryKey: [id],
-  }),
-  paginated: (params: Params = {}, filter: Filter = {}) => ({
-    queryFn: () => getBetas(params, filter),
-    queryKey: [params, filter],
-  }),
-});
 
 export const useBetasQuery = (params?: Params, filter?: Filter) =>
   useQuery<ResourcePage<Beta>, APIError[]>({
